requirements.txt
dj-database-url==0.5.0
Django==2.2.6
djangorestframework==3.10.3
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
django-cors-headers==3.2.0
